have bedpetobed12 use new stream class, but rename to avoid name collision
