+++ svtools/utils.py
