+++ tests/bedpetobed12_tests.py
@@ -10,11 +10,11 @@ class IntegrationTest_bedpetobed12(TestCase):
     def run_integration_test(self):
         test_directory = os.path.dirname(os.path.abspath(__file__))
         test_data_dir = os.path.join(test_directory, 'test_data', 'bedpetobed12')
-        input = os.path.join(test_data_dir, 'input.bed')
+        input_file = os.path.join(test_data_dir, 'input.bed')
         expected_result = os.path.join(test_data_dir, 'expected.bed')
         temp_descriptor, temp_output_path = tempfile.mkstemp(suffix='.bed')
-        with os.fdopen(temp_descriptor, 'w') as output_handle:
-            svtools.bedpetobed12.processBEDPE(input, 'BEDPE', 1000000, output_handle)
+        with open(input_file) as input_stream, os.fdopen(temp_descriptor, 'w') as output_handle:
+            svtools.bedpetobed12.processBEDPE(input_stream, 'BEDPE', 1000000, output_handle)
         expected_lines = open(expected_result).readlines()
         produced_lines = open(temp_output_path).readlines()
         diff = difflib.unified_diff(produced_lines, expected_lines, fromfile=temp_output_path, tofile=expected_result)
@@ -23,7 +23,7 @@ class IntegrationTest_bedpetobed12(TestCase):
             for line in result:
                 sys.stdout.write(line)
             self.assertFalse(result)
-        #os.remove(temp_output_path)
+        os.remove(temp_output_path)
 
 if __name__ == "__main__":
     main()
