+++ tests/util_tests.py
@@ -1,25 +1,25 @@
 from unittest import TestCase, main
 import sys
 import os
-import svtools.io.utils as sio
+import svtools.utils as su 
 
 class InputStreamTest(TestCase):
     def test_init_hyphen(self):
-        new_handle = sio.InputStream('-')
+        new_handle = su.InputStream('-')
         self.assertIs(new_handle.handle, sys.stdin)
 
     def test_valid(self):
-        self.assertTrue(sio.InputStream.valid('-'))
+        self.assertTrue(su.InputStream.valid('-'))
         with self.assertRaises(IOError):
-            sio.InputStream.valid(None)
+            su.InputStream.valid(None)
 
     def test_context_manager(self):
         test_directory = os.path.dirname(os.path.abspath(__file__))
-        test_data_dir = os.path.join(test_directory, '..', 'test_data', 'io', 'utils')
+        test_data_dir = os.path.join(test_directory, 'test_data', 'utils')
         test_input = os.path.join(test_data_dir, 'file.txt.gz')
 
         temporary_obj = None
-        with sio.InputStream(test_input) as stream:
+        with su.InputStream(test_input) as stream:
             temporary_obj = stream
             for line in stream:
                 sys.stdout.write(line)
@@ -27,10 +27,10 @@ class InputStreamTest(TestCase):
 
     def test_plain_iteration(self):
         test_directory = os.path.dirname(os.path.abspath(__file__))
-        test_data_dir = os.path.join(test_directory, '..', 'test_data', 'io', 'utils')
+        test_data_dir = os.path.join(test_directory, 'test_data', 'utils')
         test_input = os.path.join(test_data_dir, 'file.txt')
 
-        stream = sio.InputStream(test_input)
+        stream = su.InputStream(test_input)
         for line in stream:
             sys.stdout.write(line)
         stream.close()
